Update appcontrol.md
Language, logic and grammar in APP CONTROL and SMART APP CONTROL section is vague and confusing.  Smart App control being based on App control is a single thought - that is clear. What follows in the original is a complete mess and confusing and illogical, as it is impossible to determine if 'running of LOB applications' and the example policy are intended for use with App control, or with Smart App control.